mark_agent/pyproject.toml
[project]
name = "mark-agent"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "a2a-sdk>=0.3.23",
    "crewai[google-genai]>=1.9.3",
    "python-dotenv>=1.1.1",
    "sse-starlette>=3.2.0",
    "starlette>=0.52.1",
    "uvicorn>=0.41.0",
]
